added clear all hook
